Add more functionality to airline_staff.html
For the search flights, added a toggle button next to every flight
which allows airline staff to quickly change the status of any flight.

Added the 'Add Airplane' functionality, as well as a table which
displays all airplanes owned by the airline which staff works for.

Made sure that render_template() doesn't mess with the contents of
tables.