Made many security and design changes to project
Added md5 hashing through the use of a library and a helper function
which is found at the bottom of the file.

Added a wrapper function which protects airline staff page, and airline
staff functionality (e.g. search, add flights, etc.) from being used
if not in session.

Cleaned up comments, code, and whatnot for easier readability for future
of the project.

Todo:
Still have to make it so that airline staff can add their numbers. This
will involve dynamically adding insert "bubbles" to add to the
Phone_Numbers table on the html, and dealing with that in init.py file.